Initial commit for scram / pyscram source
This commit adds the initial implementation for this SCRAM
authentication library. It uses openssl for cyptography and
implements bare minimum for message handling. Onus is on
library consumer to properly process messages, validate, and
send responses. Channel binding validation is also external
to this library. A cpython extension for it is provided to
enhance testing and allow python applications to easily
implement SCRAM authentication.